Rename to InteractiveObjectTable
