+++ src/widgets/tables/LanguageTable.tsx
@@ -7,7 +7,7 @@ import HoverableEnumeration from '@features/hovercard/HoverableEnumeration';
 import HoverableObjectName from '@features/hovercard/HoverableObjectName';
 import { SortBy } from '@features/sorting/SortTypes';
 import { CodeColumn } from '@features/table/CommonColumns';
-import ObjectTable from '@features/table/ObjectTable';
+import InteractiveObjectTable from '@features/table/InteractiveObjectTable';
 import TableValueType from '@features/table/TableValueType';
 
 import { LanguageData, LanguageField } from '@entities/language/LanguageTypes';
@@ -38,7 +38,7 @@ const LanguageTable: React.FC = () => {
   };
 
   return (
-    <ObjectTable<LanguageData>
+    <InteractiveObjectTable<LanguageData>
       objects={languagesInSelectedSource}
       columns={[
         codeColumn,
