+++ src/widgets/tables/TableOfLanguagesInCensus.tsx
@@ -8,7 +8,7 @@ import { ObjectType, SearchableField } from '@features/page-params/PageParamType
 import usePageParams from '@features/page-params/usePageParams';
 import { SortBy } from '@features/sorting/SortTypes';
 import { CodeColumn } from '@features/table/CommonColumns';
-import ObjectTable from '@features/table/ObjectTable';
+import InteractiveObjectTable from '@features/table/InteractiveObjectTable';
 import TableValueType from '@features/table/TableValueType';
 
 import { CensusData } from '@entities/census/CensusTypes';
@@ -83,7 +83,7 @@ const TableOfLanguagesInCensus: React.FC<Props> = ({ census }) => {
           {langsNotFound.join(', ')}
         </div>
       )}
-      <ObjectTable<LocaleData>
+      <InteractiveObjectTable<LocaleData>
         objects={languagesInCensus}
         shouldFilterUsingSearchBar={false}
         columns={[
