+++ src/widgets/tables/TableOfLanguagesInTerritory.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import HoverableObjectName from '@features/hovercard/HoverableObjectName';
 import { SortBy } from '@features/sorting/SortTypes';
 import { CodeColumn, EndonymColumn } from '@features/table/CommonColumns';
-import ObjectTable from '@features/table/ObjectTable';
+import InteractiveObjectTable from '@features/table/InteractiveObjectTable';
 import TableValueType from '@features/table/TableValueType';
 
 import LocaleCensusCitation from '@entities/locale/LocaleCensusCitation';
@@ -29,7 +29,7 @@ const TableOfLanguagesInTerritory: React.FC<Props> = ({ territory }) => {
   }
 
   return (
-    <ObjectTable<LocaleData>
+    <InteractiveObjectTable<LocaleData>
       objects={locales}
       shouldFilterUsingSearchBar={false}
       columns={[
