+++ src/widgets/tables/TerritoryTable.tsx
@@ -5,7 +5,7 @@ import HoverableEnumeration from '@features/hovercard/HoverableEnumeration';
 import HoverableObjectName from '@features/hovercard/HoverableObjectName';
 import { SortBy } from '@features/sorting/SortTypes';
 import { CodeColumn, NameColumn } from '@features/table/CommonColumns';
-import ObjectTable from '@features/table/ObjectTable';
+import InteractiveObjectTable from '@features/table/InteractiveObjectTable';
 import TableValueType from '@features/table/TableValueType';
 
 import { getTerritoryChildren } from '@entities/lib/getObjectMiscFields';
@@ -18,7 +18,7 @@ const TerritoryTable: React.FC = () => {
   const { territories } = useDataContext();
 
   return (
-    <ObjectTable<TerritoryData>
+    <InteractiveObjectTable<TerritoryData>
       objects={territories}
       columns={[
         CodeColumn,
